+++ tests/Github.fs
@@ -9,8 +9,9 @@ open Expecto
 let tests =
     testList "Github" [
         testList "releasesSince" [
+            let release = { TagName = "1.0"; Draft = false; Prerelease = false; Body = "" }
             test "Can parse version with 'v' prefix" {
-                let githubReleases = [{ TagName = "v1.0.0"; Draft = false; Body = "" }]
+                let githubReleases = [ { release with TagName = "v1.0.0" }]
                 let expected = Standard { Version = new Version(1, 0, 0); ReleaseNotes = "" }
                 
                 let releases = releasesSince (new Version()) githubReleases
@@ -18,7 +19,7 @@ let tests =
                 Expect.equal releases.Head expected ""
             }
             test "Can parse version without 'v' prefix" {
-                let githubReleases = [{ TagName = "1.0.0"; Draft = false; Body = "" }]
+                let githubReleases = [{ release with TagName = "1.0.0" }]
                 let expected = Standard { Version = new Version(1, 0, 0); ReleaseNotes = "" }
                 
                 let releases = releasesSince (new Version()) githubReleases
@@ -26,21 +27,21 @@ let tests =
                 Expect.equal releases.Head expected ""
             }
             test "Ignores release with empty tag name" {
-                let githubReleases = [{ TagName = ""; Draft = false; Body = "" }]
+                let githubReleases = [{ release with TagName = "" }]
                 
                 let releases = releasesSince (new Version()) githubReleases
                 
                 Expect.isEmpty releases ""
             }
             test "Ignores release with null tag name" {
-                let githubReleases = [{ TagName = null; Draft = false; Body = "" }]
+                let githubReleases = [{ release with TagName = null }]
                 
                 let releases = releasesSince (new Version()) githubReleases
                 
                 Expect.isEmpty releases ""
             }
             test "Ignores release with non-version-like tag name" {
-                let githubReleases = [{ TagName = "asdf"; Draft = false; Body = "" }]
+                let githubReleases = [{ release with TagName = "asdf" }]
                 
                 let releases = releasesSince (new Version()) githubReleases
                 
@@ -48,9 +49,9 @@ let tests =
             }
             test "Ignores releases released before specified version" {
                 let githubReleases = [
-                    { TagName = "1.0"; Draft = false; Body = "" }
-                    { TagName = "2.0"; Draft = false; Body = "" }
-                    { TagName = "3.0"; Draft = false; Body = "" }
+                    { release with TagName = "1.0" }
+                    { release with TagName = "2.0" }
+                    { release with TagName = "3.0" }
                 ]
                 
                 let releases = releasesSince (new Version(2, 0)) githubReleases
@@ -59,8 +60,20 @@ let tests =
             }
             test "Ignores draft releases" {
                 let githubReleases = [
-                    { TagName = "1.0"; Draft = true; Body = "one" }
-                    { TagName = "1.0"; Draft = false; Body = "two" }
+                    { release with Draft = true; Body = "one" }
+                    { release with Draft = false; Body = "two" }
+                ]
+                let expected = Standard { Version = new Version(1, 0); ReleaseNotes = "two" }
+                
+                let releases = releasesSince (new Version()) githubReleases
+                
+                Expect.hasLength releases 1 ""
+                Expect.equal releases.Head expected ""
+            }
+            test "Ignores pre-releases" {
+                let githubReleases = [
+                    { release with Prerelease = true; Body = "one" }
+                    { release with Prerelease = false; Body = "two" }
                 ]
                 let expected = Standard { Version = new Version(1, 0); ReleaseNotes = "two" }
                 
@@ -70,14 +83,14 @@ let tests =
                 Expect.equal releases.Head expected ""
             }
             test "Is a security release when body contains security header" {
-                let githubReleases = [{ TagName = "1.0"; Draft = false; Body = "asdf ### Security asdf" }]
+                let githubReleases = [{ release with Body = "asdf ### Security asdf" }]
                 
                 let releases = releasesSince (new Version()) githubReleases
                 
                 Expect.isUnionCase releases.Head <@ Security @> ""
             }
             test "Is a standard release when body doesn't contain security header" {
-                let githubReleases = [{ TagName = "1.0"; Draft = false; Body = "asdf # Security asdf" }]
+                let githubReleases = [{ release with Body = "asdf # Security asdf" }]
                 
                 let releases = releasesSince (new Version()) githubReleases
                 
@@ -85,7 +98,7 @@ let tests =
             }
             test "Matches CVEs" {
                 let githubReleases = [
-                    { TagName = "1.0"; Draft = false; Body = "### Security CVE 2023-1234 CVE-2023-1235 CVE-2023 0000" }
+                    { release with Body = "### Security CVE 2023-1234 CVE-2023-1235 CVE-2023 0000" }
                 ]
                 
                 let releases = releasesSince (new Version()) githubReleases