+++ CHANGELOG.md
@@ -6,6 +6,16 @@
 - Log when using the fallback products directory in cases when user doesn't have permissions to write to default 
   products directory ([#177]) 
 
+- Add a workaround to get Flatpaks to be launched.  
+  You can now clear and override `$LD_LIBRARY_PATH` to allow processes defined in `processes` in your settings file to 
+  be launched using the host's libraries by setting `$MEL_LD_LIBRARY_PATH`. Mainly useful for Steam Deck users
+  (i.e. auto-launching EDMC).
+
+  `konsole` Steam launch options example: 
+  ```sh
+  LD_LIBRARY_PATH="" konsole -e env MEL_LD_LIBRARY_PATH="$LD_LIBRARY_PATH" ./MinEdLauncher %command% /autorun /autoquit
+  ```
+
 ## [0.12.2] - 2025-07-25
 
 ### Bug Fixes